use type long long

+++ src/LinAlg/hiopVectorRajaPar.cpp
@@ -1529,7 +1529,7 @@ long long hiopVectorRajaPar::numOfElemsLessThan(const double &val) const
   long long nrm = sum.get();
 
 #ifdef HIOP_USE_MPI
-  double nrm_global;
+  long long nrm_global;
   int ierr = MPI_Allreduce(&nrm, &nrm_global, 1, MPI_LONG_LONG, MPI_SUM, comm_);
   assert(MPI_SUCCESS == ierr);
   nrm = nrm_global;
@@ -1551,7 +1551,7 @@ long long hiopVectorRajaPar::numOfElemsAbsLessThan(const double &val) const
   long long nrm = sum.get();
 
 #ifdef HIOP_USE_MPI
-  double nrm_global;
+  long long nrm_global;
   int ierr = MPI_Allreduce(&nrm, &nrm_global, 1, MPI_LONG_LONG, MPI_SUM, comm_);
   assert(MPI_SUCCESS == ierr);
   nrm = nrm_global;
